misc: changed to u32 and alignment to u8 to save space

+++ include/mun_abi.h
@@ -48,9 +48,9 @@ typedef struct
     /// Type name
     const char *name;
     /// The exact size of the type in bits without any padding
-    uint64_t size_in_bits;
+    uint32_t size_in_bits;
     /// The alignment of the type
-    uint32_t alignment;
+    uint8_t alignment;
     /// Type group
     const MunTypeGroup_t group;
 } MunTypeInfo;
